server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
pyzmq>=26.0.3
httpx>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
unidecode>=1.3.8
email-validator>=2.1.0
bcrypt==3.2.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
